third_party/blink/web_tests/external/wpt/css/css-anchor-position/position-anchor-basics-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property position-anchor has initial value normal
  assert_equals: expected "normal" but got "none"
[FAIL] Property position-anchor has initial value normal in popovers
  assert_equals: expected "normal" but got "none"
Harness: the test ran to completion.
